# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/room": {
            "get": {
                "description": "Returns a paginated list of room",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "Get All Room",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseGetAllRoom"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Update an room by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "Put Update Room",
                "parameters": [
                    {
                        "description": "Put Update Room Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.RequestPostCreateUpdateRoom"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Id of room",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponsePutUpdateRoom"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/room/asset/detail/{floor}/{roomName}": {
            "get": {
                "description": "Returns a paginated list of asset that found in a room",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "Get Room Asset By Floor And Room Name (Detail)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "In which Room you want to find the asset. Type 'all' to search in all room",
                        "name": "roomName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "In which Floor you want to find the asset.",
                        "name": "floor",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseGetRoomAssetByFloorAndRoomName"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/room/asset/short/{floor}/{roomName}": {
            "get": {
                "description": "Returns a paginated list of asset that found in a room. in short format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "Get Room Asset By Floor And Room Name (Short)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "In which Room you want to find the asset. Type 'all' to search in all room",
                        "name": "roomName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "In which Floor you want to find the asset.",
                        "name": "floor",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseGetRoomAssetShortByFloorAndRoomName"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/room/most-context/{targetCol}": {
            "get": {
                "description": "Returns a list of most appear item in room by given field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "Get Most Context Room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target Column to Analyze (such as: asset_merk, asset_category, or asset_status)",
                        "name": "targetCol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseGetMostContext"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/room/{id}": {
            "delete": {
                "description": "Permanentally delete room by id",
                "tags": [
                    "Room"
                ],
                "summary": "Delete Room By Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of room",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseDeleteRoomById"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Metadata": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "entity.RequestPostCreateUpdateRoom": {
            "type": "object",
            "required": [
                "floor",
                "room_dept",
                "room_name"
            ],
            "properties": {
                "floor": {
                    "type": "string"
                },
                "room_dept": {
                    "type": "string"
                },
                "room_name": {
                    "type": "string"
                }
            }
        },
...
        "entity.ResponseDeleteRoomById": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "room deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.ResponseGetAllRoom": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Room"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "room fetched"
                },
                "metadata": {
                    "$ref": "#/definitions/entity.Metadata"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.ResponseGetMostContext": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.StatsContextTotal"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "asset fetched"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.ResponseGetRoomAssetByFloorAndRoomName": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.RoomAsset"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "room fetched"
                },
                "metadata": {
                    "$ref": "#/definitions/entity.Metadata"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "entity.ResponseGetRoomAssetShortByFloorAndRoomName": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.RoomAssetShort"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "room fetched"
                },
                "metadata": {
                    "$ref": "#/definitions/entity.Metadata"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.ResponsePutUpdateRoom": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "room updated successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.Room": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "floor": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "room_dept": {
                    "type": "string"
                },
                "room_name": {
                    "type": "string"
                }
            }
        },
        "entity.RoomAsset": {
            "type": "object",
            "properties": {
                "asset_category": {
                    "type": "string"
                },
                "asset_desc": {
                    "type": "string"
                },
                "asset_merk": {
                    "type": "string"
                },
                "asset_name": {
                    "type": "string"
                },
                "floor": {
                    "type": "string"
                },
                "room_dept": {
                    "type": "string"
                },
                "room_name": {
                    "type": "string"
                },
                "total_asset": {
                    "type": "integer"
                }
            }
        },
        "entity.RoomAssetShort": {
            "type": "object",
            "properties": {
                "asset_category": {
                    "type": "string"
                },
                "asset_name": {
                    "type": "string"
                },
                "floor": {
                    "type": "string"
                },
                "room_dept": {
                    "type": "string"
                },
                "room_name": {
                    "type": "string"
                }
            }
        },
        "entity.StatsContextTotal": {
            "type": "object",
            "properties": {
                "context": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}